project/tests/core/test_dependencies.py
def test_dependencies_edge_cases():

    # Import relevant class

    from project.core.dependencies import Dependencies

    

    # Test: Create Dependencies instance

    dependencies = Dependencies()

    assert isinstance(dependencies, Dependencies)

    

    # Test: Add invalid dependency (empty library or version)

    dependencies.add_dependency("", "1.0.0")

    assert len(dependencies.dependencies) == 0

    dependencies.add_dependency("numpy", "")

    assert len(dependencies.dependencies) == 0

    

    # Test: Add duplicate dependency

    dependencies.add_dependency("numpy", "1.21.0")

    dependencies.add_dependency("numpy", "1.21.0")

    assert len(dependencies.dependencies) == 1

    

    # Test: Remove non-existent dependency

    dependencies.remove_dependency("non_existent_library")

    assert len(dependencies.dependencies) == 1